RMII PHY: Add support for ref_clk being provided externally
RMII spec allows an alternate mode in which ref_clk is an input on both
the MAC and the PHY. This is used by a no-name "LAN8720 ETH board"
break-out board. This commit allows this mode by specifying a ref_clk_in
Subsignal in clock_pads (instead of ref_clk).